+++ src/main/java/com/epf/persistance/mapper/PlanteMapper.java
@@ -2,6 +2,8 @@ package com.epf.persistance.mapper;
 
 import com.epf.persistance.Plante;
 import org.springframework.jdbc.core.RowMapper;
+
+import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -9,13 +11,13 @@ public class PlanteMapper implements RowMapper<Plante> {
     @Override
     public Plante mapRow(ResultSet rs, int rowNum) throws SQLException {
         return new Plante(
-                rs.getInt("id_plante"),
+                rs.getLong("id_plante"),
                 rs.getString("nom"),
                 rs.getInt("point_de_vie"),
-                rs.getInt("attaque_par_seconde"),
+                rs.getBigDecimal("attaque_par_seconde"), // Correction ici
                 rs.getInt("degat_attaque"),
                 rs.getInt("cout"),
-                rs.getInt("soleil_par_seconde"),
+                rs.getBigDecimal("soleil_par_seconde"), // Correction ici
                 rs.getString("effet"),
                 rs.getString("chemin_image")
         );
